requirements.txt
PySide6~=6.8.1
pyinstaller~=6.11.1
psutil~=6.1.0
pystray~=0.19.5
pillow~=11.0.0